Get major refactor compiling
