+++ compile_flags.txt
@@ -1,0 +1,5 @@
+-DKQUEUE
+-std=c99
+-Wall
+-Wextra
+-Werror
