+++ test/integration/main.c
@@ -1,5 +1,5 @@
 #define HTTPSERVER_IMPL
-#include "../httpserver.h"
+#include "../../httpserver.h"
 #include <signal.h>
 
 #define RESPONSE "Hello, World!"
